% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diblc.R
\name{diblc}
\alias{diblc}
\title{Dynamic Improvement Bayesian Lee-Carter}
\usage{
diblc(Y, I = 3000, B = 500, df_b = 0.9)
}
\arguments{
\item{Y}{Numerical matrix with age information on rows and time information
on columns. Must be log-mortality with valid and finite values.}

\item{I}{Number of iterations for the Gibbs sampler.}

\item{B}{Number of iterations to be discarded.}

\item{df_b}{Discount factor for improvement parameters. Defaults to 0.9.}
}
\value{
A `diblc` object.
}
\description{
Performs estimation of a Bayesian Lee-Carter model for modelling mortality
date by extending the research done by Pedroza (2006) where Dynamic Linear
Models were used for performing joint inference on the model of Lee & Carter
(1992). This routine employes empirical estimates for the `alpha` parameter
and introduces a dynamic to the improvement terms, which have evolutional
variances modelled through discount factors, as suggested in West & Harrison
(1996).
This is the result of ongoing research, and this function may change with
time to reflect the latest model considered the best approach by the authors.
}
