% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diblc_filter.R
\name{diblc_filter}
\alias{diblc_filter}
\title{Dynamic Improvement Lee-Carter Model Filtering Routine}
\usage{
diblc_filter(Y, phi, delta, phi_k, df_b = 0.9)
}
\arguments{
\item{Y}{Numerical matrix with age information on rows and time information
on columns. Must be log-mortality with valid and finite values.}

\item{phi}{Observational precisions.}

\item{delta}{Trend for kappa.}

\item{phi_k}{Precision for evolution of kappa.}

\item{df_b}{Discount factor for beta errors. Defaults to 0.90.}
}
\value{
A list with online means and variances for each parameter and one
        step ahead forecasts for the observations as a `diblc_filter` object.
}
\description{
This is the filtering routine for the Dynamic Improvement Lee-Carter Model
which is estimated by the `diblc` function. See the documentation for `diblc`
for further information.
}
