% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dyn_lc.R
\name{dyn_lc}
\alias{dyn_lc}
\title{Dynamic Lee-Carter Model}
\usage{
dyn_lc(Y, I = 2510, B = 10)
}
\arguments{
\item{Y}{Numerical matrix with age information on rows and time information
on columns. Must be log-mortality with valid and finite values.}

\item{I}{Number of iterations for the Gibbs sampler.}

\item{B}{Number of iterations to be discarded.}
}
\value{
A `dyn_lc` object.
}
\description{
Performs estimation of a fully dybamic Lee-Carter model with a gibbs estimator for
estimation of variances.
}
