% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dyn_lc_fixed.R
\name{dyn_lc_fixed}
\alias{dyn_lc_fixed}
\title{Dynamic Lee-Carter Model with Fixed Precisions}
\usage{
dyn_lc_fixed(Y, phi = NULL, phi_a = NULL, phi_b = NULL, phi_k = NULL,
  phi_d = NULL)
}
\arguments{
\item{Y}{Numerical matrix with age information on rows and time information
on columns. Must be log-mortality with valid and finite values.}

\item{phi}{Observational precisions.}

\item{phi_a}{Evolutional precisions for alpha.}

\item{phi_b}{Evolutional precisions for beta.}

\item{phi_k}{Evolutional precision for kappa.}

\item{phi_d}{Evolutional precision for delta.}
}
\value{
A list with online means and variances for each parameter and
        forecasts. Class `dyn_lc_fixed`.
}
\description{
Fully-dynamic extension of the Lee-Carter model but using fixed precisions
for observational _and_ evolutional errors.
}
