% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dyn_lc_semifixed.R
\name{dyn_lc_semifixed}
\alias{dyn_lc_semifixed}
\title{Dynamic Lee-Carter Model with Fixed Observational Precisions}
\usage{
dyn_lc_semifixed(Y, phi = NULL, df_a = 0.99, df_b = 0.9, df_k = 0.9,
  df_d = 0.99, inf_a = T)
}
\arguments{
\item{Y}{Numerical matrix with age information on rows and time information
on columns. Must be log-mortality with valid and finite values.}

\item{phi}{Observational precisions.}

\item{df_a}{Discount factor for alpha errors.}

\item{df_b}{Discount factor for beta errors.}

\item{df_k}{Discount factor for kappa and delta errors.}

\item{inf_a}{Controls if a prior based on the data should be used.
Note that this is heresy.}
}
\value{
A list with online means and variances for each parameter and
        forecasts. Same class as `dyn_lc_fixed`.
}
\description{
Fully-dynamic extension of the Lee-Carter model but using fixed precisions
for observational errors and discount factors for evolutional errors.
}
