% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/old_diblc_2.R
\name{old_diblc_2}
\alias{old_diblc_2}
\title{Dynamic Improvement Bayesian Lee-Carter}
\usage{
old_diblc_2(Y, I = 3000, B = 500)
}
\arguments{
\item{Y}{Numerical matrix with age information on rows and time information
on columns. Must be log-mortality with valid and finite values.}

\item{I}{Number of iterations for the Gibbs sampler.}

\item{B}{Number of iterations to be discarded.}
}
\value{
A `diblc` object.
}
\description{
This routine is kept for archiving reasons and making old methodology
available. See `diblc` documentation for the latest method.
}
\details{
This routine in particular considers empirical estimates of `alpha` and
estimates individual evolutional variances for each of the state-space
parameters.

This routines gives too much freedom for the evolutional variances of the
improvement terms.
}
