% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/old_diblc_filter_1.R
\name{old_diblc_filter_1}
\alias{old_diblc_filter_1}
\title{Dynamic Improvement Bayesian Lee-Carter Model}
\usage{
old_diblc_filter_1(Y, phi = NULL, df_a = 0.99, df_b = 0.9,
  df_k = 0.9, df_d = 0.99)
}
\arguments{
\item{Y}{Numerical matrix with age information on rows and time information
on columns. Must be log-mortality with valid and finite values.}

\item{phi}{Observational precisions.}

\item{df_a}{Discount factor for alpha errors. Defaults to 0.99.}

\item{df_b}{Discount factor for beta errors. Defaults to 0.90.}

\item{df_k}{Discount factor for kappa error. Defaults to 0.90.}

\item{df_d}{Discount factor for delta error. Defaults to 0.99.}
}
\value{
A list with online means and variances for each parameter and
        forecasts. Same class as `old_diblc_filter_1`.
}
\description{
See `old_diblc_1` for further documentation.
}
