% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/old_diblc_filter_2.R
\name{old_diblc_filter_2}
\alias{old_diblc_filter_2}
\title{Dynamic Improvement Bayesian Lee-Carter Model}
\usage{
old_diblc_filter_2(Y, phi, delta, phi_k, phi_b)
}
\arguments{
\item{Y}{Numerical matrix with age information on rows and time information
on columns. Must be log-mortality with valid and finite values.}

\item{phi}{Observational precisions.}

\item{delta}{Trend for kappa.}

\item{phi_k}{Precision for evolution of kappa.}

\item{phi_b}{Precision for evolution of improvement terms.}
}
\value{
A list with online means and variances for each parameter and
        forecasts as a `diblc_filter` object.
}
\description{
See `old_diblc_1` for further documentation.
}
