% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dyn_lc_methods.R
\name{plot.dyn_lc}
\alias{plot.dyn_lc}
\title{Displays graphical summaries for `dyn_lc_filter` results.}
\usage{
\method{plot}{dyn_lc}(x, y, at = NULL, ages = NULL, years = NULL,
  cred = 0.95, age = NULL, ...)
}
\arguments{
\item{x}{The `dyn_lc` result.}

\item{y}{The parameter of interest.}

\item{at}{If applicable, the time point of interest.}

\item{ages}{Vector with the ages if plotting 'beta' or 'alpha'. Optional.}

\item{years}{Vector with the years if plotting 'kappa'. Optional.}

\item{cred}{The credibility for the interval. Defaults to 0.95.}

\item{age}{The age _index_ to check one-step ahead forecasts.}

\item{...}{Unused.}
}
\value{
A `ggplot2` object.
}
\description{
Displays graphical summaries for `dyn_lc_filter` results.
}
