% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dyn_lc_fixed_methods.R
\name{plot.dyn_lc_fixed}
\alias{plot.dyn_lc_fixed}
\title{Displays graphical summaries for `dyn_lc_fixed` results.}
\usage{
\method{plot}{dyn_lc_fixed}(x, y, at = NULL, ages = NULL, years = NULL,
  cred = 0.95, ...)
}
\arguments{
\item{x}{The `dyn_lc_fixed` result.}

\item{y}{The parameter of interest.}

\item{at}{If applicable, the time point of interest.}

\item{ages}{Vector with the ages if plotting 'beta' or 'alpha'. Optional.}

\item{years}{Vector with the years if plotting 'kappa'. Optional.}

\item{cred}{The credibility for the interval. Defaults to 0.95.}
}
\value{
A `ggplot2` object.
}
\description{
Displays graphical summaries for `dyn_lc_fixed` results.
}
