% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dyn_lc_filter.R
\name{dyn_lc_filter}
\alias{dyn_lc_filter}
\title{Dynamic Lee-Carter Model Filtering Routine}
\usage{
dyn_lc_filter(Y, phi = NULL, df_a = 0.99, df_b = 0.9, df_k = 0.9,
  df_d = 0.99)
}
\arguments{
\item{Y}{Numerical matrix with age information on rows and time information
on columns. Must be log-mortality with valid and finite values.}

\item{phi}{Observational precisions.}

\item{df_a}{Discount factor for alpha errors. Defaults to 0.99.}

\item{df_b}{Discount factor for beta errors. Defaults to 0.90.}

\item{df_k}{Discount factor for kappa error. Defaults to 0.90.}

\item{df_d}{Discount factor for delta error. Defaults to 0.99.}
}
\value{
A list with online means and variances for each parameter and
        forecasts. Same class as `dyn_lc_filter`.
}
\description{
Fully-dynamic extension of the Lee-Carter model but using fixed precisions
for observational errors and discount factors for evolutional errors.
}
